# backend/tests/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AddressCreate": {
        "properties": {
          "country_code": {
            "const": "KE",
            "default": "KE",
            "title": "Country Code",
            "type": "string"
          },
          "county": {
            "maxLength": 80,
            "minLength": 1,
            "title": "County",
            "type": "string"
          },
          "delivery_notes": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Notes"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "label": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,7}0*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "line_1": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Line 1",
            "type": "string"
          },
          "line_2": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line 2"
          },
          "locality": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Locality",
            "type": "string"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,7}0*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "postal_code": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          }
        },
        "required": [
          "label",
          "line_1",
          "locality",
          "county"
        ],
        "title": "AddressCreate",
        "type": "object"
      },
      "AddressResponse": {
        "properties": {
          "country_code": {
            "const": "KE",
            "default": "KE",
            "title": "Country Code",
            "type": "string"
          },
          "county": {
            "maxLength": 80,
            "minLength": 1,
            "title": "County",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delivery_notes": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Notes"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "label": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "latitude": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,7}0*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "line_1": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Line 1",
            "type": "string"
          },
          "line_2": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line 2"
          },
          "locality": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Locality",
            "type": "string"
          },
          "longitude": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,7}0*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "postal_code": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "label",
          "line_1",
          "locality",
          "county",
          "id",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "AddressResponse",
        "type": "object"
      },
      "AddressUpdate": {
        "properties": {
          "county": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "County"
          },
          "delivery_notes": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Notes"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,7}0*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "line_1": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line 1"
          },
          "line_2": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line 2"
          },
          "locality": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locality"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,7}0*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "postal_code": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          }
        },
        "title": "AddressUpdate",
        "type": "object"
      },
      "BuyerProfileResponse": {
        "properties": {
          "buyer_type": {
            "title": "Buyer Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "organization_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Name"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "organization_name",
          "buyer_type",
          "created_at",
          "updated_at"
        ],
        "title": "BuyerProfileResponse",
        "type": "object"
      },
      "BuyerProfileUpdate": {
        "properties": {
          "buyer_type": {
            "anyOf": [
              {
                "enum": [
                  "individual",
                  "business",
                  "institution"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer Type"
          },
          "organization_name": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Name"
          }
        },
        "title": "BuyerProfileUpdate",
        "type": "object"
      },
...
      "FarmerProfileResponse": {
        "properties": {
          "county": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "County"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delivery_radius_km": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Radius Km"
          },
          "farm_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farm Name"
          },
          "farm_size_acres": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farm Size Acres"
          },
          "farming_practices": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farming Practices"
          },
          "farming_since_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farming Since Year"
          },
          "locality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locality"
          },
          "offers_delivery": {
            "default": false,
            "title": "Offers Delivery",
            "type": "boolean"
          },
          "offers_pickup": {
            "default": true,
            "title": "Offers Pickup",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "verification_status": {
            "title": "Verification Status",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "farm_name",
          "county",
          "verification_status",
          "created_at",
          "updated_at"
        ],
        "title": "FarmerProfileResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InstallationUpsert": {
        "properties": {
          "platform": {
            "enum": [
              "android",
              "ios",
              "web"
            ],
            "title": "Platform",
            "type": "string"
          },
          "push_token": {
            "anyOf": [
              {
                "maxLength": 4096,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Push Token"
          }
        },
        "required": [
          "platform"
        ],
        "title": "InstallationUpsert",
        "type": "object"
      },
...
      "ProfileResponse": {
        "properties": {
          "avatar_object_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Object Key"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "display_name",
          "avatar_object_key",
          "locale",
          "bio",
          "created_at",
          "updated_at"
        ],
        "title": "ProfileResponse",
        "type": "object"
      },
      "ProfileUpdate": {
        "properties": {
          "avatar_object_key": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Object Key"
          },
          "bio": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "locale": {
            "anyOf": [
              {
                "maxLength": 16,
                "pattern": "^[a-z]{2}(?:-[A-Z]{2})?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          }
        },
        "title": "ProfileUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Mavuno API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "get_profile_api_v1_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Profile",
        "tags": [
          "profiles"
        ]
      },
      "patch": {
        "operationId": "update_profile_api_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Profile",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/users/me/addresses": {
      "get": {
        "operationId": "list_addresses_api_v1_users_me_addresses_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AddressResponse"
                  },
                  "title": "Response List Addresses Api V1 Users Me Addresses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Addresses",
        "tags": [
          "profiles"
        ]
      },
      "post": {
        "operationId": "create_address_api_v1_users_me_addresses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Address",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/users/me/addresses/{address_id}": {
      "delete": {
        "operationId": "delete_address_api_v1_users_me_addresses__address_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "address_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Address Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Address",
        "tags": [
          "profiles"
        ]
      },
      "patch": {
        "operationId": "update_address_api_v1_users_me_addresses__address_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "address_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Address Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Address",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/users/me/buyer": {
      "patch": {
        "operationId": "update_buyer_profile_api_v1_users_me_buyer_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Buyer Profile",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/users/me/farmer": {
      "get": {
        "operationId": "get_farmer_profile_api_v1_users_me_farmer_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Farmer Profile",
        "tags": [
          "profiles"
        ]
      },
      "patch": {
        "operationId": "update_farmer_profile_api_v1_users_me_farmer_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FarmerProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Farmer Profile",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/users/me/installations/{installation_id}": {
      "delete": {
        "operationId": "revoke_installation_api_v1_users_me_installations__installation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "installation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Installation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Installation",
        "tags": [
          "profiles"
        ]
      },
      "put": {
        "operationId": "register_installation_api_v1_users_me_installations__installation_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "installation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Installation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallationUpsert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register Installation",
        "tags": [
          "profiles"
        ]
      }
    },
...
  }
}